Rename: 메인 박스 속성값 코드 컨밴션 반영하여 수정
- num => number
- img => profileImage
- rel => relationship
- date => createdAt
- name => recipientId
