+++ src/components/Home/MainBoxSecond.jsx
@@ -10,7 +10,7 @@ export default function MainBoxSecond() {
         <img src={EmojiMainImage} alt="이모지" />
       </div>
       <MainTextBox
-        num="02"
+        number="02"
         text="서로에게 이모지로 감정을 표현해보세요"
         subText="롤링 페이퍼에 이모지를 추가할 수 있어요."
       />
