--- src/components/Home/MainImgBox.jsx
@@ -1,5 +1,0 @@
-import * as S from './MainImgItemBox.style';
-
-export default function MainImgBox() {
-  return <S.ItemBox>jj</S.ItemBox>;
-}
