+++ src/components/Home/MainItemBasicBox.jsx
@@ -1,4 +1,4 @@
-import * as S from './MainImgItemBox.style';
+import * as S from './MainItemBox.style';
 
 export default function MainItemBasicBox() {
   return <S.ItemBox>jj</S.ItemBox>;
