+++ src/components/Home/MainTextBox.jsx
@@ -1,9 +1,9 @@
 import * as S from './MainTextBox.style';
 
-export default function MainTextBox({ num, text, subText }) {
+export default function MainTextBox({ number, text, subText }) {
   return (
     <S.TextContainer>
-      <S.PointBox>Point.{num}</S.PointBox>
+      <S.PointBox>Point.{number}</S.PointBox>
       <S.TextBox>
         <S.Text>{text}</S.Text>
         <S.SubText>{subText}</S.SubText>
